requirements.txt
absl-py==0.11.0
aiofiles==0.6.0
appdirs==1.4.4
appnope @ file:///opt/concourse/worker/volumes/live/5f13e5b3-5355-4541-5fc3-f08850c73cf9/volume/appnope_1606859448618/work
argon2-cffi @ file:///opt/concourse/worker/volumes/live/59af29ac-4890-416e-7ab7-794f8d6f7ecd/volume/argon2-cffi_1596828548321/work
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
cachetools==4.1.1
certifi==2019.11.28
cffi @ file:///opt/concourse/worker/volumes/live/730e9a28-66f9-4e03-51ad-252ec8e40d81/volume/cffi_1606255126408/work
chardet==3.0.4
Click==7.0
cmake==3.18.4.post1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dlib==19.21.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.63.0
fastapi-jwt-auth==0.5.0
filelock==3.0.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.34.0
h11==0.12.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///opt/concourse/worker/volumes/live/88f541d3-5a27-498f-7391-f2e50ca36560/volume/ipykernel_1596206680118/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/26969e8f-c9f7-42dc-6ffb-b3effd424c49/volume/ipython_1604101242376/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort==5.5.5
jedi @ file:///opt/concourse/worker/volumes/live/bd2e4eb2-df59-4bd6-4802-01c9ab8f36aa/volume/jedi_1608920706165/work
Jinja2==2.11.2
joblib==1.0.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///opt/concourse/worker/volumes/live/e8302867-5dbe-440b-7a37-f61bacc38ed8/volume/jupyter_core_1606148993907/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
labelImg==1.8.4
lazr.uri==1.0.5
lazy-object-proxy==1.4.3
lxml==4.6.2
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe @ file:///opt/concourse/worker/volumes/live/cb778296-98db-45ad-411e-6f726e102dc3/volume/markupsafe_1594371638608/work
matplotlib==3.3.3
mccabe==0.6.1
mistune @ file:///opt/concourse/worker/volumes/live/95802d64-d39c-491b-74ce-b9326880ca54/volume/mistune_1594373201816/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///opt/concourse/worker/volumes/live/2b9c1d93-d0fd-432f-7d93-66c93d81b614/volume/nbconvert_1601914875037/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
netifaces==0.10.4
networkx==2.5
notebook @ file:///opt/concourse/worker/volumes/live/d85c0794-1357-4fff-61be-761a5e60b85b/volume/notebook_1608835566278/work
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.0
pandocfilters @ file:///opt/concourse/worker/volumes/live/c330e404-216d-466b-5327-8ce8fe854d3a/volume/pandocfilters_1605120442288/work
parso==0.7.0
pexpect==4.6.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==7.0.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.3
Pygments @ file:///tmp/build/80754af9/pygments_1607368905949/work
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyngrok==5.0.2
pyparsing==2.4.7
PyQt5==5.14.1
PyQt5-sip==12.8.1
pyRFC3339==1.1
pyrsistent @ file:///opt/concourse/worker/volumes/live/ff11f3f0-615b-4508-471d-4d9f19fa6657/volume/pyrsistent_1600141727281/work
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.4
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
Shapely==1.7.1
six==1.14.0
sklearn==0.0
starlette==0.13.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.11.26
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado @ file:///opt/concourse/worker/volumes/live/05341796-4198-4ded-4a9a-332fde3cdfd1/volume/tornado_1606942323372/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
urllib3==1.25.8
uvicorn==0.13.3
virtualenv==20.0.33
wadllib==1.3.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xmljson==0.2.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
